fix: harden measurement qualification report
