+++ tests/lib/measurement-qualification.test.ts
@@ -3,8 +3,10 @@ import { describe, expect, it } from 'vitest';
 import {
   buildMeasurementQualificationReport,
   createAsiaBangkokMeasurementWindow,
+  createMeasurementQualificationStore,
   type MeasurementEventRow,
   type MeasurementPaymentRow,
+  type MeasurementQualificationQueries,
   type MeasurementWebVitalRow,
 } from '@/lib/measurement-qualification';
 
@@ -25,7 +27,7 @@ function payment(overrides: Partial<MeasurementPaymentRow> = {}): MeasurementPay
     currency: 'THB',
     method: 'stripe',
     status: 'completed',
-    createdAt: new Date('2026-08-20T03:00:00.000Z'),
+    purchaseCompletedAt: new Date('2026-08-20T03:00:00.000Z'),
     ...overrides,
   };
 }
@@ -45,6 +47,7 @@ function event(overrides: Partial<MeasurementEventRow> = {}): MeasurementEventRo
     learningEnrollmentId: null,
     lessonId: null,
     source: 'server',
+    metadata: JSON.stringify({ method: 'stripe' }),
     privacyUnsafe: false,
     createdAt: new Date('2026-08-20T03:01:00.000Z'),
     ...overrides,
@@ -61,6 +64,7 @@ function exposure(index: number): MeasurementEventRow {
     userId: null,
     paymentId: null,
     source: 'client',
+    metadata: JSON.stringify({ placement: 'course_detail' }),
   });
 }
 
@@ -69,7 +73,11 @@ function vital(
   deviceClass: 'mobile' | 'desktop',
   index: number,
 ): MeasurementWebVitalRow {
-  const value = metricName === 'LCP' ? 2_000 + index : metricName === 'INP' ? 150 + index : 0.08 + index / 1_000;
+  const value = metricName === 'LCP'
+    ? 2_000 + index
+    : metricName === 'INP'
+      ? 120 + index
+      : 0.02 + index / 1_000;
   return {
     pageLoadId: `page-${metricName}-${deviceClass}-${index}`,
     metricName,
@@ -78,7 +86,7 @@ function vital(
     releaseIdentity: 'release-1',
     value: String(value),
     rating: 'good',
-    updatedAt: new Date('2026-08-20T04:00:00.000Z'),
+    createdAt: new Date('2026-08-20T04:00:00.000Z'),
   };
 }
 
@@ -123,6 +131,43 @@ describe('measurement qualification report', () => {
     expect(report.qualification.gates.reconciliation.passed).toBe(false);
   });
 
+  it('anchors purchase eligibility to the committed transition and reports later refunds separately', () => {
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [
+        payment({
+          id: 'completed-before-window',
+          purchaseCompletedAt: new Date('2026-08-17T16:59:59.000Z'),
+        }),
+        payment(),
+        payment({
+          id: 'refunded-2',
+          attributedExposureId: null,
+          status: 'refunded',
+          purchaseCompletedAt: new Date('2026-08-21T03:00:00.000Z'),
+        }),
+      ],
+      events: [
+        event(),
+        event({
+          id: 'fact-refunded-2',
+          paymentId: 'refunded-2',
+          attributedExposureId: null,
+          createdAt: new Date('2026-08-21T03:01:00.000Z'),
+        }),
+      ],
+      webVitals: [],
+    });
+
+    expect(report.purchaseReconciliation).toMatchObject({
+      eligiblePaymentCount: 2,
+      reconciledPaymentCount: 2,
+    });
+    expect(report.diagnostics.refundedPaymentCount).toBe(1);
+  });
+
   it('reports attributed conversion separately from operational unattributed trends', () => {
     const report = buildMeasurementQualificationReport({
       window,
@@ -155,24 +200,28 @@ describe('measurement qualification report', () => {
         id: 'workspace-1', eventName: 'learning_workspace_started', exposureId: 'workspace-exposure-1',
         attributedExposureId: null, userId: null, courseId: 'course-1', paymentId: null,
         learningEnrollmentId: 'enrollment-1', lessonId: 'lesson-1', source: 'client',
+        metadata: null,
         createdAt: new Date('2026-08-20T00:00:00.000Z'),
       }),
       event({
         id: 'lesson-fact-1', eventName: 'lesson_completed', attributedExposureId: null,
         userId: null, paymentId: null, learningFactId: 'progress-1',
         learningEnrollmentId: 'enrollment-1', lessonId: 'lesson-1',
+        metadata: null,
         createdAt: new Date('2026-08-23T00:00:00.000Z'),
       }),
       event({
         id: 'course-fact-1', eventName: 'course_completed', attributedExposureId: null,
         userId: null, paymentId: null, learningFactId: 'enrollment-1',
         learningEnrollmentId: 'enrollment-1', lessonId: null,
+        metadata: null,
         createdAt: new Date('2026-08-24T00:00:00.000Z'),
       }),
       event({
         id: 'workspace-2', eventName: 'learning_workspace_started', exposureId: 'workspace-exposure-2',
         attributedExposureId: null, userId: null, courseId: 'course-1', paymentId: null,
         learningEnrollmentId: 'enrollment-2', lessonId: 'lesson-2', source: 'client',
+        metadata: null,
         createdAt: new Date('2026-08-30T00:00:00.000Z'),
       }),
     ];
@@ -193,6 +242,48 @@ describe('measurement qualification report', () => {
       nextLessonCompletedWithin7DaysCount: 1,
       courseCompletedWithin7DaysCount: 1,
       continuityRate: 1,
+    });
+  });
+
+  it('evaluates every workspace start but deduplicates learner outcomes by enrollment', () => {
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [],
+      events: [
+        event({
+          id: 'review-start', eventName: 'learning_workspace_started',
+          exposureId: 'review-exposure', attributedExposureId: null, userId: null,
+          paymentId: null, learningEnrollmentId: 'enrollment-1', lessonId: 'lesson-review',
+          source: 'client', metadata: null,
+          createdAt: new Date('2026-08-20T00:00:00.000Z'),
+        }),
+        event({
+          id: 'qualified-start', eventName: 'learning_workspace_started',
+          exposureId: 'qualified-exposure', attributedExposureId: null, userId: null,
+          paymentId: null, learningEnrollmentId: 'enrollment-1', lessonId: 'lesson-2',
+          source: 'client', metadata: null,
+          createdAt: new Date('2026-08-22T00:00:00.000Z'),
+        }),
+        event({
+          id: 'qualified-completion', eventName: 'lesson_completed',
+          attributedExposureId: null, userId: null, paymentId: null,
+          learningFactId: 'progress-2', learningEnrollmentId: 'enrollment-1',
+          lessonId: 'lesson-2', metadata: null,
+          createdAt: new Date('2026-08-23T00:00:00.000Z'),
+        }),
+      ],
+      webVitals: [],
+    });
+
+    expect(report.learner).toMatchObject({
+      workspaceStartObservationCount: 2,
+      startedEnrollmentCount: 1,
+      maturedEnrollmentCount: 1,
+      nextLessonCompletedWithin7DaysCount: 1,
+      continuityRate: 1,
+      medianHoursToNextLessonCompletion: 24,
     });
   });
 
@@ -240,6 +331,21 @@ describe('measurement qualification report', () => {
     ))).toBe(true);
   });
 
+  it('keeps Web Vitals membership on the immutable first-observed timestamp', () => {
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [],
+      events: [],
+      webVitals: [{ ...vital('LCP', 'mobile', 1), createdAt: window.end }],
+      minimumWebVitalSampleCount: 1,
+    });
+
+    expect(report.webVitals.releaseIdentities).toEqual([]);
+    expect(report.diagnostics.invalidWebVitalCount).toBe(1);
+  });
+
   it('rejects baseline readiness when schema, eligibility, privacy, release, or reconciliation fails', () => {
     const report = buildMeasurementQualificationReport({
       window,
@@ -257,5 +363,126 @@ describe('measurement qualification report', () => {
 
     expect(report.qualification.status).toBe('rejected');
     expect(Object.values(report.qualification.gates).every((gate) => !gate.passed)).toBe(true);
+  });
+
+  it('rejects untrusted release sentinels even when the release identity is unique', () => {
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [],
+      events: [],
+      webVitals: [{ ...vital('LCP', 'mobile', 1), releaseIdentity: 'unknown-release' }],
+      minimumWebVitalSampleCount: 1,
+    });
+
+    expect(report.qualification.gates.release).toEqual({
+      passed: false,
+      reasons: ['release_identity_untrusted'],
+    });
+  });
+
+  it('rejects malformed or privacy-bearing metadata without echoing its contents', () => {
+    const privateMetadata = JSON.stringify({ method: 'stripe', email: 'learner@example.test' });
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [payment()],
+      events: [event({ metadata: privateMetadata })],
+      webVitals: [],
+    });
+
+    expect(report.qualification.gates.schema.passed).toBe(false);
+    expect(report.qualification.gates.privacy.passed).toBe(false);
+    expect(report.diagnostics.invalidMetadataCount).toBe(1);
+    expect(JSON.stringify(report)).not.toContain('learner@example.test');
+  });
+
+  it('caps reconciliation identities while preserving the total count', () => {
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: Array.from({ length: 55 }, (_, index) => payment({ id: `payment-${index}` })),
+      events: [],
+      webVitals: [],
+    });
+
+    expect(report.purchaseReconciliation.missing).toMatchObject({
+      count: 55,
+      truncated: true,
+    });
+    expect(report.purchaseReconciliation.missing.identities).toHaveLength(50);
+  });
+
+  it('qualifies a complete baseline that passes every gate', () => {
+    const webVitals = (['LCP', 'INP', 'CLS'] as const).flatMap((metricName) => (
+      (['mobile', 'desktop'] as const).flatMap((deviceClass) => (
+        Array.from({ length: 75 }, (_, index) => vital(metricName, deviceClass, index))
+      ))
+    ));
+    const report = buildMeasurementQualificationReport({
+      window,
+      observedAt,
+      control: approvedControl,
+      payments: [payment()],
+      events: [...Array.from({ length: 100 }, (_, index) => exposure(index + 1)), event()],
+      webVitals,
+    });
+
+    expect(report.qualification.status).toBe('qualified');
+    expect(Object.values(report.qualification.gates).every((entry) => entry.passed)).toBe(true);
+  });
+
+  it('merges baseline and learner-tail queries with exact bounds and deduplicated identities', async () => {
+    const observedWindows: Record<string, { start: Date; end: Date }> = {};
+    const sharedFact = event({ id: 'shared-fact', privacyUnsafe: false });
+    const tailCompletion = event({
+      id: 'tail-completion',
+      eventName: 'course_completed',
+      attributedExposureId: null,
+      userId: null,
+      paymentId: null,
+      learningFactId: 'enrollment-1',
+      learningEnrollmentId: 'enrollment-1',
+      lessonId: null,
+      metadata: null,
+      createdAt: new Date('2026-09-02T00:00:00.000Z'),
+    });
+    const queries: MeasurementQualificationQueries = {
+      async readPayments(queryWindow) {
+        observedWindows.payments = queryWindow;
+        return [payment()];
+      },
+      async readBaselineEvents(queryWindow) {
+        observedWindows.baseline = queryWindow;
+        return [{ ...sharedFact, privacyUnsafe: 1 }];
+      },
+      async readLearnerTailEvents(queryWindow) {
+        observedWindows.tail = queryWindow;
+        return [{ ...tailCompletion, privacyUnsafe: 0 }];
+      },
+      async readLinkedPurchaseEvents(queryWindow, observedBefore) {
+        observedWindows.linked = queryWindow;
+        expect(observedBefore).toEqual(observedAt);
+        return [{ ...sharedFact, privacyUnsafe: 1 }];
+      },
+      async readWebVitals(queryWindow) {
+        observedWindows.vitals = queryWindow;
+        return [];
+      },
+    };
+
+    const snapshot = await createMeasurementQualificationStore(queries).readSnapshot(window, observedAt);
+
+    expect(observedWindows.baseline).toEqual(window);
+    expect(observedWindows.tail).toEqual({
+      start: window.end,
+      end: new Date('2026-09-07T17:00:00.000Z'),
+    });
+    expect(observedWindows.linked).toEqual(window);
+    expect(snapshot.events.map((row) => row.id)).toEqual(['shared-fact', 'tail-completion']);
+    expect(snapshot.events[0]?.privacyUnsafe).toBe(true);
   });
 });
